+++ vm.h
@@ -30,14 +30,14 @@ typedef struct vm_ports {
 
 typedef struct vm_state {
 	vm_core *cores;
-	uint16_t core_count;
+	uint8_t core_count;
 
 	vm_ports ports;
 	uint8_t memory[UINT16_MAX];
 } vm_state;
 
-void vm_init(vm_state *, uint16_t core_count, vm_core *cores);
-void vm_step(vm_state *, uint16_t core_index);
+void vm_init(vm_state *, uint8_t core_count, vm_core *cores);
+void vm_step(vm_state *, uint8_t core_index);
 
 char const *vm_padded_reg_name(uint8_t);
 
@@ -48,7 +48,7 @@ char const *vm_fault_name(vm_fault);
 
 // these point to mutable static memory
 char const *vm_disasm(uint8_t, uint8_t, uint8_t);
-char const *vm_disasm_pc(vm_state const *, uint16_t core_index);
+char const *vm_disasm_pc(vm_state const *, uint8_t core_index);
 
 
 #endif // VM_H
